De-duplicate ClutLerp
