+++ src/st/nz0/4070C.c
@@ -424,42 +424,7 @@ INCLUDE_ASM("st/nz0/nonmatchings/4070C", func_801C1AD8);
 
 INCLUDE_ASM("st/nz0/nonmatchings/4070C", func_801C20B8);
 
-void ClutLerp(RECT* rect, u16 palIdxA, u16 palIdxB, s32 steps, u16 offset) {
-    u16 buf[COLORS_PER_PAL];
-    RECT bufRect;
-    s32 factor;
-    u32 t;
-    u32 r, g, b;
-    s32 i, j;
-    u16 *palA, *palB;
-
-    bufRect.x = rect->x;
-    bufRect.w = COLORS_PER_PAL;
-    bufRect.h = 1;
-
-    palA = g_Clut + palIdxA * COLORS_PER_PAL;
-    palB = g_Clut + palIdxB * COLORS_PER_PAL;
-
-    for (i = 0; i < steps; i++) {
-        factor = i * 4096 / steps;
-        for (j = 0; j < COLORS_PER_PAL; j++) {
-            r = (palA[j] & 0x1F) * (4096 - factor) + (palB[j] & 0x1F) * factor;
-            g = ((palA[j] >> 5) & 0x1F) * (4096 - factor) +
-                ((palB[j] >> 5) & 0x1F) * factor;
-            b = ((palA[j] >> 10) & 0x1F) * (4096 - factor) +
-                ((palB[j] >> 10) & 0x1F) * factor;
-
-            t = palA[j] & 0x8000;
-            t |= palB[j] & 0x8000;
-
-            buf[j] = t | (r >> 12) | ((g >> 12) << 5) | ((b >> 12) << 10);
-        }
-
-        bufRect.y = rect->y + i;
-        LoadImage(&bufRect, buf);
-        g_ClutIds[offset + i] = GetClut(bufRect.x, bufRect.y);
-    }
-}
+#include "../clut_lerp.h"
 
 void func_801C29B0(s16 sfxId) {
     s32 var_a3;
